+++ src/MailHide.php
@@ -1,15 +1,17 @@
 <?php
+
 /**
- * @copyright Copyright (c) 2005-2015 Zend Technologies USA Inc. (http://www.zend.com)
- * @license   http://framework.zend.com/license/new-bsd New BSD License
+ * @see       https://github.com/laminas/laminas-recaptcha for the canonical source repository
+ * @copyright https://github.com/laminas/laminas-recaptcha/blob/master/COPYRIGHT.md
+ * @license   https://github.com/laminas/laminas-recaptcha/blob/master/LICENSE.md New BSD License
  */
 
-namespace ZendService\ReCaptcha;
+namespace Laminas\ReCaptcha;
 
+use Laminas\Stdlib\ArrayUtils;
+use Laminas\Validator\EmailAddress as EmailAddressValidator;
+use Laminas\Validator\ValidatorInterface;
 use Traversable;
-use Zend\Stdlib\ArrayUtils;
-use Zend\Validator\EmailAddress as EmailAddressValidator;
-use Zend\Validator\ValidatorInterface;
 
 /**
  * Render and validate MailHide reCaptchas.