+++ src/ReCaptcha.php
@@ -1,15 +1,17 @@
 <?php
+
 /**
- * @copyright Copyright (c) 2005-2015 Zend Technologies USA Inc. (http://www.zend.com)
- * @license   http://framework.zend.com/license/new-bsd New BSD License
+ * @see       https://github.com/laminas/laminas-recaptcha for the canonical source repository
+ * @copyright https://github.com/laminas/laminas-recaptcha/blob/master/COPYRIGHT.md
+ * @license   https://github.com/laminas/laminas-recaptcha/blob/master/LICENSE.md New BSD License
  */
-namespace ZendService\ReCaptcha;
+namespace Laminas\ReCaptcha;
 
 use Exception as PhpException;
+use Laminas\Http\Client as HttpClient;
+use Laminas\Http\Request as HttpRequest;
+use Laminas\Stdlib\ArrayUtils;
 use Traversable;
-use Zend\Http\Client as HttpClient;
-use Zend\Http\Request as HttpRequest;
-use Zend\Stdlib\ArrayUtils;
 
 /**
  * Render and verify ReCaptchas
@@ -434,7 +436,7 @@ HTML;
      * Gets a solution to the verify server
      *
      * @param string $responseField
-     * @return \Zend\Http\Response
+     * @return \Laminas\Http\Response
      * @throws Exception
      */
     protected function post($responseField)
@@ -469,10 +471,10 @@ HTML;
      * Verify the user input
      *
      * This method calls up the post method and returns a
-     * \ZendService\ReCaptcha\Response object.
+     * \Laminas\ReCaptcha\Response object.
      *
      * @param string $responseField
-     * @return \ZendService\ReCaptcha\Response
+     * @return \Laminas\ReCaptcha\Response
      */
     public function verify($responseField)
     {